wip procassembly converter for proc16a

+++ tools/Assembler/Assembly/AssemblyUtil.cs
@@ -3,11 +3,11 @@ using Assembler.Parsing.ProcAssemblyV2;
 namespace Assembler.Assembly;
 
 public static class AssemblyUtil {
-	public static int SetBit(int instruction, int bit, bool value) {
+	public static ushort SetBit(ushort instruction, int bit, bool value) {
 		if (value) {
-			return instruction | (1 << bit);
+			return (ushort) (instruction | (1 << bit));
 		} else {
-			return instruction & (~(1 << bit));
+			return (ushort) (instruction & (~(1 << bit)));
 		}
 	}
 }
